DAOs 2.

+++ src/main/java/com/librarySystem/dao/GenericDAO.java
@@ -8,9 +8,9 @@ public interface GenericDAO<T> {
 
 	public void update(int index, T t);
 
-	public void remove(int index);
+	public void remove(long index);
 
-	public T getByIndex(int index);
+	public T getByIndex(long index);
 
 	public List<T> getAll(Class<T> t);
